Added a cast suggested by Thomas Heilgenmann <thomas.heiligenmann@klinik-lowwenstein.de> to make warnings go away on the Alpha.
CVS patchset: 2028
CVS date: 2003/06/18 17:11:48

+++ lib/clplumbing/ipcsocket.c
@@ -800,7 +800,7 @@ socket_resume_io_write(struct IPC_CHANNEL *ch, gboolean* started)
 
 		if (sendrc != msg->msg_len) {
 			cl_perror("Sent %d byte message body: rc = %d"
-			,	msg->msg_len, sendrc);
+			,	(int)msg->msg_len, sendrc);
 		}
 
 #if 0
